pyproject.toml
[project]
name = "vortexserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aioredis>=2.0.1",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqladmin>=0.21.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
